feat: refactor Media3 session logic and implement media item resolution
